+++ ObjectPool/Specialized/IStringBuilderPool.cs
@@ -1,4 +1,4 @@
-﻿// File name: StringBuilderPool.cs
+﻿// File name: IStringBuilderPool.cs
 //
 // Author(s): Alessio Parma <alessio.parma@gmail.com>
 //
@@ -26,26 +26,9 @@ using System.Text;
 namespace CodeProject.ObjectPool.Specialized
 {
     /// <summary>
-    ///   An <see cref="IObjectPool{PooledStringBuilder}"/> ready to be used.
-    ///   <see cref="StringBuilder"/> management can be further configured using the
-    ///   <see cref="MinimumStringBuilderCapacity"/> and <see cref="MaximumStringBuilderCapacity"/> properties.
+    ///   An object pool specialized in <see cref="StringBuilder"/> management.
     /// </summary>
-    public sealed class StringBuilderPool : ObjectPool<PooledStringBuilder>, IStringBuilderPool
+    public interface IStringBuilderPool : IObjectPool<PooledStringBuilder>
     {
-        /// <summary>
-        ///   Thread-safe pool instance.
-        /// </summary>
-        public static IStringBuilderPool Instance { get; } = new StringBuilderPool();
-
-        /// <summary>
-        ///   Minimum capacity a <see cref="StringBuilder"/> should have when created. Defaults to 4 * 1024 characters.
-        /// </summary>
-        public static int MinimumStringBuilderCapacity { get; set; } = 4 * 1024;
-
-        /// <summary>
-        ///   Maximum capacity a <see cref="StringBuilder"/> might have in order to be able to return
-        ///   to pool. Defaults to 512 * 1024 characters.
-        /// </summary>
-        public static int MaximumStringBuilderCapacity { get; set; } = 512 * 1024;
     }
 }