+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed default min and max size for MemoryStreamPool: 4KB min, 512KB max.
 * Changed default min and max size for StringBuilderPool: 4K char min, 512K char max.
+* Created two ad-hoc interfaces for specialized pools.
 
 ### v2.0.5 (2016-08-23) ###